+++ src/components/CardPreview.tsx
